Reuses concatenated alg if present

+++ nprlib/task/concat_alg.py
@@ -6,7 +6,7 @@ log = logging.getLogger("main")
 from nprlib.task import Msf
 from nprlib.master_task import ConcatAlgTask
 from nprlib.master_job import Job
-from nprlib.utils import SeqGroup, GLOBALS, generate_runid, strip
+from nprlib.utils import SeqGroup, GLOBALS, generate_runid, strip, pexist
 from nprlib import db
 from nprlib.errors import TaskError
 
@@ -59,32 +59,36 @@ class ConcatAlg(ConcatAlgTask):
     def finish(self):
         # Assumes tasks resulting from genetree workflow, in which
         # only Alg and Acleaner tasks could contain the results
-        for job in self.jobs: 
-            if job.ttype == "alg" and job.nodeid not in self.job2alg:
-                self.job2alg[job.nodeid] = job.alg_fasta_file
-            elif job.ttype == "acleaner":
-                self.job2alg[job.nodeid] = job.clean_alg_fasta_file
-            elif job.ttype == "mchooser":
-                self.job2model[job.nodeid] = job.best_model
-        if self.cog_ids - set(self.job2alg):
-            log.error("Missing %s algs", len(self.cog_ids -
-                                             set(self.job2alg)))
-            raise TaskError(self)
+        if pexist(self.alg_fasta_file) and pexist(self.alg_phylip_file) \
+                and pexist(self.partitions_file):
+            log.log(28, "@@5:Concatenated alignment is already present. Skipping...@@1:")
+        else: 
+            for job in self.jobs: 
+                if job.ttype == "alg" and job.nodeid not in self.job2alg:
+                    self.job2alg[job.nodeid] = job.alg_fasta_file
+                elif job.ttype == "acleaner":
+                    self.job2alg[job.nodeid] = job.clean_alg_fasta_file
+                elif job.ttype == "mchooser":
+                    self.job2model[job.nodeid] = job.best_model
+            if self.cog_ids - set(self.job2alg):
+                log.error("Missing %s algs", len(self.cog_ids -
+                                                 set(self.job2alg)))
+                raise TaskError(self)
 
-        alg_data = [(self.job2alg[nid], self.job2model.get(nid, self.default_model)) for nid in self.job2alg]
-        filenames, models = zip(*alg_data)
+            alg_data = [(self.job2alg[nid], self.job2model.get(nid, self.default_model)) for nid in self.job2alg]
+            filenames, models = zip(*alg_data)
 
-        mainalg, partitions, sp2alg, species = get_concatenated_alg(filenames,
-                                    models, sp_field=0,
-                                    sp_delimiter=GLOBALS["spname_delimiter"])
-        
-        log.log(20, "Done concat alg, now writting fasta")
-        mainalg.write(outfile=self.alg_fasta_file, format="fasta")
-        log.log(20, "Done concat alg, now writting phylip")
-        mainalg.write(outfile=self.alg_phylip_file, format="iphylip_relaxed")
-        open(self.partitions_file, "w").write('\n'.join(partitions))
-        log.log(20, "Done concat alg")
-        log.log(26, "Modeled regions: \n"+'\n'.join(partitions))
+            mainalg, partitions, sp2alg, species = get_concatenated_alg(filenames,
+                                                                        models, sp_field=0,
+                                                                        sp_delimiter=GLOBALS["spname_delimiter"])
+
+            log.log(20, "Done concat alg, now writting fasta")
+            mainalg.write(outfile=self.alg_fasta_file, format="fasta")
+            log.log(20, "Done concat alg, now writting phylip")
+            mainalg.write(outfile=self.alg_phylip_file, format="iphylip_relaxed")
+            open(self.partitions_file, "w").write('\n'.join(partitions))
+            log.log(20, "Done concat alg")
+            log.log(26, "Modeled regions: \n"+'\n'.join(partitions))
                         
 def get_species_code(name, splitter, field):
     # By default, taxid is the first par of the seqid, separated by
